Add online meetings (#6572)
* Add config to allow online meetings

* Add online_meeting_url fields to DB

* Add fields to forms and commands

* Use `data-` attributes instead of specific field names

* Remove unused locales

* Fix JS

* Fix locale

* Fix forms specs

* Fix specs

* Hide meeting type and online url fields if online meetings are not allowed

* Show online meeting URL

* Ensure fields are hidden

* Fix meeting form

* Fix tests

* Add `type_of`meeting` column to meetings

It stores whether the meeting is in-person or online

* Filter meetings by type

* Add missing i18n string

* Validate URLs

* Address feedback

* Fix i18n

* Remove config option

* test: select the type of meeting before filling in location

* Move URL details to cell

* Fix form

* Fix icon

* Fix edit form
