+++ decidim-meetings/app/assets/config/decidim_meetings_manifest.js
@@ -3,6 +3,7 @@
 //= link decidim/meetings/admin/registrations_form.js
 //= link decidim/meetings/admin/registrations_invite_form.js
 //= link decidim/meetings/admin/meetings_form.js
+//= link decidim/meetings/meetings_form.js
 //= link decidim/meetings/admin/agendas.js
 //= link decidim/meetings/admin/destroy_meeting_alert.js
 //= link_tree ../images/decidim
