+++ decidim-meetings/app/cells/decidim/meetings/meeting_m/address.erb
@@ -1,1 +1,5 @@
-<%= cell "decidim/address", model %>
+<% if online_meeting? %>
+  <%= cell "decidim/meetings/meeting_url", model %>
+<% else %>
+  <%= cell "decidim/address", model %>
+<% end %>
