+++ decidim-meetings/app/cells/decidim/meetings/meeting_m_cell.rb
@@ -27,6 +27,8 @@ module Decidim
         decidim_html_escape(present(model).title)
       end
 
+      delegate :online_meeting?, to: :model
+
       private
 
       def spans_multiple_dates?
