+++ decidim-meetings/app/cells/decidim/meetings/meeting_url/show.erb
@@ -1,0 +1,15 @@
+<div class="card__icondata">
+  <ul class="card-data">
+    <li class="card-data__item">
+      <%= resource_icon %>
+    </li>
+    <li class="card-data__item">
+      <div class="text-left">
+        <div class="card__icondata--address">
+          <span><%= link_to(model.online_meeting_url, model.online_meeting_url, target: "_blank") %></span><br>
+          <span><%= translated_attribute model.location_hints %></span>
+        </div>
+      </div>
+    </li>
+  </ul>
+</div>
