+++ decidim-meetings/app/cells/decidim/meetings/meeting_url_cell.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Decidim
+  module Meetings
+    # This cell renders the URL address of an online meeting.
+    class MeetingUrlCell < Decidim::ViewModel
+      include Cell::ViewModel::Partial
+      include LayoutHelper
+
+      private
+
+      def resource_icon
+        icon "video", class: "icon--big", role: "img", "aria-hidden": true
+      end
+    end
+  end
+end
