+++ decidim-meetings/app/commands/decidim/meetings/create_meeting.rb
@@ -46,6 +46,8 @@ module Decidim
           author: form.current_user,
           decidim_user_group_id: form.user_group_id,
           registration_terms: { I18n.locale => form.current_component.settings.default_registration_terms },
+          online_meeting_url: form.online_meeting_url,
+          type_of_meeting: form.clean_type_of_meeting,
           component: form.current_component
         }
