+++ decidim-meetings/app/commands/decidim/meetings/update_meeting.rb
@@ -55,7 +55,9 @@ module Decidim
             location: { I18n.locale => form.location },
             location_hints: { I18n.locale => form.location_hints },
             author: form.current_user,
-            decidim_user_group_id: form.user_group_id
+            decidim_user_group_id: form.user_group_id,
+            type_of_meeting: form.clean_type_of_meeting,
+            online_meeting_url: form.online_meeting_url
           },
           visibility: "public-only"
         )
