+++ decidim-meetings/app/types/decidim/meetings/meeting_type.rb
@@ -60,6 +60,8 @@ module Decidim
           [meeting.latitude, meeting.longitude]
         }
       end
+      field :typeOfMeeting, !types.String, "The type of the meeting (online or in-person)"
+      field :onlineMeetingUrl, !types.String, "The URL of the meeting (when the type is online)"
     end
   end
 end
