+++ decidim-meetings/app/views/decidim/meetings/admin/meetings/new.html.erb
@@ -1,4 +1,4 @@
-<%= decidim_form_for(@form, html: { class: "form new_meeting" }) do |f| %>
+<%= decidim_form_for(@form, html: { class: "form meetings_form new_meeting" }) do |f| %>
   <%= render partial: "form", object: f, locals: { title: t(".title") } %>
 
   <div class="button--double form-general-submit">
