+++ decidim-meetings/app/views/decidim/meetings/meetings/_form.html.erb
@@ -7,11 +7,18 @@
 </div>
 
 <div class="field">
-  <%= form.geocoding_field :address %>
+  <%= form.select :type_of_meeting,
+                  @form.type_of_meeting_select,
+                  { include_blank: t(".select_a_meeting_type") },
+                  { multiple: false } %>
+</div>
+
+<div class="field" data-meeting-type="in_person">
+  <%= form.geocoding_field :address, input_html: { required: false } %>
   <p class="help-text"><%= t(".address_help") %></p>
 </div>
 
-<div class="field">
+<div class="field" data-meeting-type="in_person">
   <%= form.text_area :location %>
   <p class="help-text"><%= t(".location_help") %></p>
 </div>
@@ -19,6 +26,11 @@
 <div class="field">
   <%= form.text_area :location_hints %>
   <p class="help-text"><%= t(".location_hints_help") %></p>
+</div>
+
+<div class="field" data-meeting-type="online">
+  <%= form.text_field :online_meeting_url %>
+  <p class="help-text"><%= t(".online_meeting_url_help") %></p>
 </div>
 
 <div class="field">
