+++ decidim-meetings/app/views/decidim/meetings/meetings/show.html.erb
@@ -119,7 +119,13 @@ edit_link(
         </span>
       <% end %>
       <%= decidim_sanitize(present(meeting).description(links: true)) %>
-      <%= render partial: "decidim/shared/static_map", locals: { icon_name: "meetings", geolocalizable: meeting } %>
+
+      <% if meeting.online_meeting? %>
+        <%= render partial: "online_meeting_link", locals: { meeting: meeting } %>
+      <% else %>
+        <%= render partial: "decidim/shared/static_map", locals: { icon_name: "meetings", geolocalizable: meeting } %>
+      <% end %>
+
       <%= render partial: "decidim/shared/tags", locals: { resource: meeting, tags_class_extra: "tags--meeting" } %>
     </div>
