+++ decidim-meetings/db/migrate/20201006140511_add_online_meeting_url.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddOnlineMeetingUrl < ActiveRecord::Migration[5.2]
+  def change
+    add_column :decidim_meetings_meetings, :online_meeting_url, :string
+  end
+end
