+++ decidim-meetings/db/migrate/20201009124057_add_type_of_meeting.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddTypeOfMeeting < ActiveRecord::Migration[5.2]
+  def change
+    add_column :decidim_meetings_meetings, :type_of_meeting, :string, default: "in_person"
+  end
+end
