+++ decidim-meetings/lib/decidim/meetings/test/factories.rb
@@ -34,10 +34,16 @@ FactoryBot.define do
     transparent { true }
     questionnaire { build(:questionnaire) }
     registration_form_enabled { true }
+    type_of_meeting { :in_person }
     component { build(:component, manifest_name: "meetings") }
 
     author do
       component.try(:organization)
+    end
+
+    trait :online do
+      type_of_meeting { :online }
+      online_meeting_url { "https://decidim.org" }
     end
 
     trait :official do
