+++ decidim-meetings/spec/commands/admin/update_meeting_spec.rb
@@ -24,6 +24,8 @@ module Decidim::Meetings
     let(:user) { create :user, :admin, organization: organization }
     let(:private_meeting) { false }
     let(:transparent) { true }
+    let(:type_of_meeting) { "online" }
+    let(:online_meeting_url) { "http://decidim.org" }
     let(:form) do
       double(
         invalid?: invalid,
@@ -42,7 +44,9 @@ module Decidim::Meetings
         transparent: transparent,
         services_to_persist: services_to_persist,
         current_user: user,
-        current_organization: organization
+        current_organization: organization,
+        clean_type_of_meeting: type_of_meeting,
+        online_meeting_url: online_meeting_url
       )
     end
 
@@ -124,7 +128,9 @@ module Decidim::Meetings
             transparent: transparent,
             services_to_persist: services_to_persist,
             current_user: user,
-            current_organization: organization
+            current_organization: organization,
+            clean_type_of_meeting: type_of_meeting,
+            online_meeting_url: online_meeting_url
           )
         end
