+++ decidim-meetings/spec/forms/admin/meeting_form_spec.rb
@@ -49,6 +49,8 @@ module Decidim::Meetings
     let(:category_id) { category.id }
     let(:private_meeting) { false }
     let(:transparent) { true }
+    let(:type_of_meeting) { "in_person" }
+    let(:online_meeting_url) { "http://decidim.org" }
     let(:attributes) do
       {
         decidim_scope_id: scope_id,
@@ -63,7 +65,9 @@ module Decidim::Meetings
         end_time: end_time,
         private_meeting: private_meeting,
         transparent: transparent,
-        services: services_attributes
+        services: services_attributes,
+        type_of_meeting: type_of_meeting,
+        online_meeting_url: online_meeting_url
       }
     end
 
@@ -87,7 +91,8 @@ module Decidim::Meetings
       it { is_expected.not_to be_valid }
     end
 
-    describe "when location is missing" do
+    describe "when location is missing and type of meeting is in_person" do
+      let(:type_of_meeting) { "in_person" }
       let(:location) { { en: nil } }
 
       it { is_expected.not_to be_valid }
@@ -176,6 +181,19 @@ module Decidim::Meetings
       subject { form.number_of_services }
 
       it { is_expected.to eq(services.size) }
+    end
+
+    describe "when online meeting link is missing and type of meeting is online" do
+      let(:type_of_meeting) { "online" }
+      let(:online_meeting_url) { nil }
+
+      it { is_expected.not_to be_valid }
+    end
+
+    describe "when type of meeting is missing" do
+      let(:type_of_meeting) { nil }
+
+      it { is_expected.not_to be_valid }
     end
   end
 end
