+++ decidim-meetings/spec/system/user_creates_meeting_spec.rb
@@ -49,6 +49,7 @@ describe "User creates meeting", type: :system do
         let(:longitude) { 2.1234 }
         let!(:meeting_start_time) { Time.current + 2.days }
         let(:meeting_end_time) { meeting_start_time + 4.hours }
+        let(:online_meeting_url) { "http://decidim.org" }
         let(:meeting_scope) { create :scope, organization: organization }
         let(:datetime_format) { I18n.t("time.formats.decidim_short") }
         let(:time_format) { I18n.t("time.formats.time_of_day") }
@@ -77,6 +78,7 @@ describe "User creates meeting", type: :system do
           within ".new_meeting" do
             fill_in :meeting_title, with: meeting_title
             fill_in :meeting_description, with: meeting_description
+            select "In person", from: :meeting_type_of_meeting
             fill_in :meeting_location, with: meeting_location
             fill_in :meeting_location_hints, with: meeting_location_hints
             fill_in_geocoding :meeting_address, with: meeting_address
@@ -115,6 +117,7 @@ describe "User creates meeting", type: :system do
               within ".new_meeting" do
                 fill_in :meeting_title, with: meeting_title
                 fill_in :meeting_description, with: meeting_description
+                select "In person", from: :meeting_type_of_meeting
                 fill_in :meeting_location, with: meeting_location
                 fill_in :meeting_location_hints, with: meeting_location_hints
                 fill_in :meeting_start_time, with: meeting_start_time.strftime(datetime_format)
@@ -137,6 +140,7 @@ describe "User creates meeting", type: :system do
             within ".new_meeting" do
               fill_in :meeting_title, with: meeting_title
               fill_in :meeting_description, with: meeting_description
+              select "In person", from: :meeting_type_of_meeting
               fill_in :meeting_location, with: meeting_location
               fill_in :meeting_location_hints, with: meeting_location_hints
               fill_in_geocoding :meeting_address, with: meeting_address
@@ -179,6 +183,24 @@ describe "User creates meeting", type: :system do
             click_link "New meeting"
             expect(page).to have_selector("#authorizationModal")
             expect(page).to have_content("Authorization required")
+          end
+        end
+
+        it "lets the user choose the meeting type" do
+          visit_component
+
+          click_link "New meeting"
+
+          within ".new_meeting" do
+            select "In person", from: :meeting_type_of_meeting
+            expect(page).to have_field("Address")
+            expect(page).to have_field(:meeting_location)
+            expect(page).to have_no_field("Online meeting URL")
+
+            select "Online", from: :meeting_type_of_meeting
+            expect(page).to have_no_field("Address")
+            expect(page).to have_no_field(:meeting_location)
+            expect(page).to have_field("Online meeting URL")
           end
         end
       end
